Proper symbolic link handling part (2)

+++ urbackupclient/clientdao.cpp
@@ -25,7 +25,8 @@ const int ClientDAO::c_is_group = 0;
 const int ClientDAO::c_is_user = 1;
 const int ClientDAO::c_is_system_user = 2;
 
-ClientDAO::ClientDAO(IDatabase *pDB)
+ClientDAO::ClientDAO(IDatabase *pDB, bool with_files_tmp)
+	: with_files_tmp(with_files_tmp)
 {
 	db=pDB;
 	prepareQueries();
@@ -36,10 +37,19 @@ ClientDAO::~ClientDAO()
 	destroyQueries();
 }
 
-void ClientDAO::prepareQueries(void)
+void ClientDAO::prepareQueries()
 {
 	q_get_files=db->Prepare("SELECT data,num FROM files WHERE name=?", false);
-	q_add_files=db->Prepare("INSERT INTO files_tmp (name, num, data) VALUES (?,?,?)", false);
+
+	if(with_files_tmp)
+	{
+		q_add_files=db->Prepare("INSERT INTO files_tmp (name, num, data) VALUES (?,?,?)", false);
+	}
+	else
+	{
+		q_add_files=NULL;
+	}
+	
 	q_get_dirs=db->Prepare("SELECT name, path, id, optional, tgroup, symlinked FROM backupdirs", false);
 	q_remove_all=db->Prepare("DELETE FROM files", false);
 	q_get_changed_dirs=db->Prepare("SELECT id, name FROM mdirs WHERE name GLOB ? UNION SELECT id, name FROM mdirs_backup WHERE name GLOB ?", false);
@@ -51,8 +61,17 @@ void ClientDAO::prepareQueries(void)
 	q_remove_shadowcopies=db->Prepare("DELETE FROM shadowcopies WHERE id=?", false);
 	q_save_changed_dirs=db->Prepare("INSERT OR REPLACE INTO mdirs_backup SELECT id, name FROM mdirs WHERE name GLOB ?", false);
 	q_delete_saved_changed_dirs=db->Prepare("DELETE FROM mdirs_backup", false);
-	q_copy_from_tmp_files=db->Prepare("INSERT INTO files (num, data, name) SELECT num, data, name FROM files_tmp", false);
-	q_delete_tmp_files=db->Prepare("DELETE FROM files_tmp", false);
+	if(with_files_tmp)
+	{
+		q_copy_from_tmp_files=db->Prepare("INSERT INTO files (num, data, name) SELECT num, data, name FROM files_tmp", false);
+		q_delete_tmp_files=db->Prepare("DELETE FROM files_tmp", false);
+	}
+	else
+	{
+		q_copy_from_tmp_files=NULL;
+		q_delete_tmp_files=NULL;
+	}
+	
 	q_has_changed_gap=db->Prepare("SELECT name FROM mdirs WHERE name GLOB '##-GAP-##*'", false);
 	q_get_del_dirs=db->Prepare("SELECT name FROM del_dirs WHERE name GLOB ? UNION SELECT name FROM del_dirs_backup WHERE name GLOB ?", false);
 	q_del_del_dirs=db->Prepare("DELETE FROM del_dirs WHERE name GLOB ?", false);
@@ -192,9 +211,13 @@ bool ClientDAO::getFiles(std::wstring path, std::vector<SFileAndHash> &data)
 		{
 			memcpy(&ss, ptr, sizeof(unsigned short));
 			ptr+=sizeof(unsigned short);
-			tmp.resize(ss);
-			memcpy(&tmp[0], ptr, ss);
-			f.symlink_target=Server->ConvertToUnicode(tmp);
+			if(ss>0)
+			{
+				tmp.resize(ss);
+				memcpy(&tmp[0], ptr, ss);
+				f.symlink_target=Server->ConvertToUnicode(tmp);
+				ptr+=ss;
+			}			
 		}
 
 		data.push_back(f);
